+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Generate `typedef struct <iface>VTable <iface>VTable;` and `typedef struct <iface> <iface>;` prior to their definitions to allow accepting `<ifaceVTable>` and `<iface>` as interface function parameters.
 
+### Changed
+
+ - Generate `char dummy;` only for an empty virtual table (i.e., a marker interface without superinterfaces).
+
 ## [0.1.0] - 2021-06-23
 
 ### Added